ForbesTools/requirements.txt
appdirs==1.4.0
awsebcli==3.9.0
backports.ssl-match-hostname==3.5.0.1
blessed==1.9.5
botocore==1.5.7
cement==2.8.2
click==6.7
colorama==0.3.7
Django==1.10.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
httplib2==0.10.3
jmespath==0.9.1
oauth2==1.9.0.post1
packaging==16.8
pathspec==0.5.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-dotenv==0.6.3
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
wcwidth==0.1.7
websocket-client==0.40.0
